refactor: move getPreview to tika_preview.go, rename has_preview.go
