+++ services/search/pkg/content/tika.go
@@ -6,7 +6,6 @@ import (
 	"io"
 	"net/http"
 	"slices"
-	"strconv"
 	"strings"
 
 	gateway "github.com/cs3org/go-cs3apis/cs3/gateway/v1beta1"
@@ -16,7 +15,6 @@ import (
 
 	"github.com/opencloud-eu/opencloud/pkg/log"
 	"github.com/opencloud-eu/opencloud/services/search/pkg/config"
-	"github.com/opencloud-eu/opencloud/services/thumbnails/pkg/thumbnail"
 )
 
 // Tika is used to extract content from a resource,
@@ -152,43 +150,4 @@ func (t Tika) detectLanguage(ctx context.Context, content string) string {
 		}
 	}
 	return ""
-}
-
-// frontCoverDescription is the picture type Tika reports (as dc:description) for
-// the front cover, matching the thumbnailer's selection in the dhowden/tag fork.
-const frontCoverDescription = "Cover (front)"
-
-// getPreview returns the dimensions of an audio file's embedded cover art from
-// Tika's recursive metadata, preferring the front cover (dc:description) and
-// falling back to the first image, matching the thumbnailer's selection. It only
-// runs for EmbeddedPreviewMimeTypes.
-func getPreview(mimeType string, metas []map[string][]string) *Preview {
-	if _, ok := thumbnail.EmbeddedPreviewMimeTypes[mimeType]; !ok {
-		return nil
-	}
-	var first *Preview
-	for _, meta := range metas {
-		ct, err := getFirstValue(meta, "Content-Type")
-		if err != nil || !strings.HasPrefix(ct, "image/") {
-			continue
-		}
-		w, wErr := getFirstValue(meta, "tiff:ImageWidth")
-		h, hErr := getFirstValue(meta, "tiff:ImageLength")
-		if wErr != nil || hErr != nil {
-			continue
-		}
-		width, wErr := strconv.ParseInt(w, 10, 32)
-		height, hErr := strconv.ParseInt(h, 10, 32)
-		if wErr != nil || hErr != nil || width <= 0 || height <= 0 {
-			continue
-		}
-		preview := &Preview{Width: int32(width), Height: int32(height)}
-		if desc, _ := getFirstValue(meta, "dc:description"); desc == frontCoverDescription {
-			return preview
-		}
-		if first == nil {
-			first = preview
-		}
-	}
-	return first
 }
