+++ services/thumbnails/pkg/thumbnail/has_preview.go
