+++ services/thumbnails/pkg/thumbnail/has_preview_test.go
